used APIRouter to organize user routes
