+++ sql_app/routers/__init__.py
@@ -1,0 +1,1 @@
+from .users_routes import user_routes
